# Conflicts:
#	src/main/java/com/cb/contractfarming/agromerchants/Merchant.java
#	src/main/java/com/cb/contractfarming/agromerchants/service/MerchantRepository.java
#	src/main/java/com/cb/contractfarming/agromerchants/service/MerchantService.java

+++ src/main/java/com/cb/contractfarming/agromerchants/service/MerchantService.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 public interface MerchantService {
     Merchant register(Merchant merchant);
 
-    Optional<Merchant> findById(int id);
+    Optional<Merchant> findById(Long id);
 
     List<Merchant> findByName(String name);
